定时器重构
