+++ go.mod
@@ -2,7 +2,4 @@ module github.com/Quieting/corn
 
 go 1.12
 
-require (
-	github.com/bwmarrin/snowflake v0.3.0
-	github.com/robfig/cron/v3 v3.0.0
-)
+require github.com/bwmarrin/snowflake v0.3.0
